requirements.txt
# Core PyTorch dependencies (as specified in README)
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Python version requirement
# python>=3.9

# Basic dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
pyyaml>=6.0

# Deep learning and ML
transformers>=4.20.0
h5py>=3.7.0
tensorboard>=2.9.0
tensorboardx>=2.5

# Audio processing
librosa>=0.9.0
soundfile>=0.10.0

# Video processing
opencv-python>=4.6.0
imageio>=2.19.0
Pillow>=9.0.0

# Text processing
nltk>=3.7
spacy>=3.4.0

# Graph neural networks (for graph fusion)
torch-geometric>=2.1.0
torch-scatter>=2.0.9
torch-sparse>=0.6.15
torch-cluster>=1.6.0

# Utilities
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Web API (optional for deployment)
flask>=2.2.0
flask-cors>=3.0.10
requests>=2.28.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.15.0
